gSOAP 2.8.55 stable
git-svn-id: https://svn.code.sf.net/p/gsoap2/code@130 06ac4634-b970-40ce-84dc-3611c6d995e3

+++ gsoap/src/lex.yy.c
@@ -812,7 +812,7 @@ static int hexchar(int*);
 static int octchar(int*);
 static void module(const char *name, const char *fullname);
 static void import(const char *file);
-static int magic(const char *name);
+static long magic(const char *name);
 #define YY_NO_INPUT 1
 
 #line 819 "lex.yy.c"
@@ -3243,9 +3243,9 @@ static void module(const char *name, const char *fullname)
   }
 }
 
-static int magic(const char *name)
+static long magic(const char *name)
 { size_t i;
-  int n;
+  long n;
   if (strlen(name) > 4)
     semerror("#module name length must not exceed four characters");
   n = 0;
